separate the server rendering and client rendering
