+++ app/not-found.tsx
@@ -2,7 +2,7 @@ export default function NotFound() {
   return (
     <div>
       <h2>找不到相關資訊</h2>
-      <p>Could not find requested resource</p>
+      <p>Could not find the requested resource</p>
     </div>
   );
 }
