+++ app/order/404/page.tsx
@@ -1,0 +1,3 @@
+export default function NotFound() {
+  return <div>查無此訂單</div>;
+}
